Separate all routes into own domain files
