+++ backend/src/main/kotlin/no/bekk/database/DatabaseComment.kt
@@ -1,0 +1,12 @@
+package no.bekk.database
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DatabaseComment(
+    val actor: String,
+    val questionId: String,
+    val comment: String,
+    val team: String?,
+    val updated: String
+)
